+++ testdata/de/dm/intellij/liferay/language/osgi/run/GogoShellLineMarkerProviderTest/MyOsgiCommand.java
@@ -1,0 +1,10 @@
+import org.osgi.service.component.annotations.Component;
+
+@Component(
+		property = {"osgi.command.scope=foo", "osgi.command.function=bar"},
+		service = Object.class
+)
+public class MyOsgiCommand {
+	public void bar()  {
+	}
+}