+++ changelog.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Feature: add inspection for  `setUserId()` should not be used in `update` methods in `LocalServiceImpl` classes (avoid unexpected permission expansions)
 - Feature: add syntax highlighting, bracket matching and folding for OSGi `.config` files
 - Feature: add references for `osgi.command.function` property used for Gogo Shell commands
+- Feature: Run Configuration for gogo shell commands
 - Bugfix: improve compatibility with IntelliJ Community Edition
 
 ## 0.0.25